refactor pipeline code, multi-branch support for replay pipeline, show diffs in auto-executor CI report
- Code cleanups and refactors
- Support multi brach runs for different java versions and branches, support docker starts and non docker starts
- Autoexecutor CI report now shows operation type, expected and actual values for failing cases.
- adding ./idea to .gitignore

+++ src/test/python/replay_target.py
@@ -1,13 +1,19 @@
 import os
 import sys
-from Target import Target, ReplayTest
-from configEnum import buildSystem, TestResult, ReportType
+from Target import Target, ReplayTest, TargetRunProperties, ReplayTestOptions
+from configEnum import buildSystem, TestResult, ReportType, StartMode
 from markup_report_generator import Report_Generator
 import subprocess
 
 def replay_target(target):
     # clone repo
-    os.system("git clone " + target.test_repo_url)
+    print(f"Starting replay test run for {target.test_repo_name} -> branch {target.target_run_properties.branch_name}")
+    os.system(f"git clone -b {target.target_run_properties.branch_name} {target.test_repo_url}")
+
+    # set java version
+    expected_java_version = target.target_run_properties.java_version
+    target.set_java_home(f"/usr/lib/jvm/temurin-{expected_java_version}-jdk-amd64")
+    target.check_java_version(expected_java_version)
 
     # modify build
     if (target.buildSystem == buildSystem.MAVEN):
@@ -16,19 +22,27 @@ def replay_target(target):
         target.modify_gradle(sdk_version)
 
     # server start
-    docker_up_cmd = "cd " + target.test_repo_name + " && docker compose -f conf/docker-compose.yml up -d"
-    val_1 = os.system(docker_up_cmd)
+    if target.target_run_properties.start_mode == StartMode.DOCKER:
+        docker_up_cmd = "cd " + target.test_repo_name + " && docker compose -f conf/docker-compose.yml up -d"
+        val_1 = os.system(docker_up_cmd)
 
-    proc = subprocess.Popen(["docker ps -a"], stdout=subprocess.PIPE, shell=True)
-    (out_stream, err_stream) = proc.communicate()
-    docker_container_name = "target-repo"
+        proc = subprocess.Popen(["docker ps -a"], stdout=subprocess.PIPE, shell=True)
+        (out_stream, err_stream) = proc.communicate()
+        docker_container_name = "target-repo"
 
-    # target replay
-    if (target.buildSystem == buildSystem.MAVEN):
-        test_command = "docker exec " + docker_container_name + " ./mvnw clean install surefire:test --fail-never"
-    elif (target.buildSystem == buildSystem.GRADLE):
-        test_command = "docker exec " + docker_container_name + " ./gradlew test"
-    val_2 = os.system(test_command)
+        # target replay
+        if (target.buildSystem == buildSystem.MAVEN):
+            test_command = "docker exec " + docker_container_name + " ./mvnw clean install surefire:test --fail-never"
+        elif (target.buildSystem == buildSystem.GRADLE):
+            test_command = "docker exec " + docker_container_name + " ./gradlew test"
+        val_2 = os.system(test_command)
+    else:
+        if (target.buildSystem == buildSystem.MAVEN):
+            test_command = "cd "+target.test_repo_name+" && ./mvnw clean install surefire:test --fail-never"
+        elif (target.buildSystem == buildSystem.GRADLE):
+            test_command = "cd "+target.test_repo_name+" && ./gradlew test"
+        val_1 = 0
+        val_2 = os.system(test_command)
 
     response_code = val_1 or val_2
     if (response_code == 0):
@@ -38,8 +52,10 @@ def replay_target(target):
 
     # assert and clean repo
     result_map = target.check_replay()
-    docker_down_cmd = "cd " + target.test_repo_name + " && docker compose -f conf/docker-compose.yml down"
-    os.system(docker_down_cmd)
+    if target.target_run_properties.start_mode == StartMode.DOCKER:
+        docker_down_cmd = "cd " + target.test_repo_name + " && docker compose -f conf/docker-compose.yml down"
+        os.system(docker_down_cmd)
+
     os.system("rm -rf " + target.test_repo_name)
     return result_map
 
@@ -54,6 +70,8 @@ if __name__ == "__main__":
             "/pom.xml",
             "/src/main/java/org/unlogged/demo/UnloggedDemoApplication.java",
             buildSystem.MAVEN,
+            target_run_properties = TargetRunProperties("main","17", StartMode.DOCKER),
+            replay_test_options= ReplayTestOptions(
             [
                 ReplayTest("org.unlogged.demo.controller.CustomerController.saveCustomerProfile - normal",
                            TestResult.PASS),
@@ -328,8 +346,8 @@ if __name__ == "__main__":
                 ReplayTest("PropertyServiceCEImpl.updateExisting - normal", TestResult.PASS),
                 ReplayTest("org.unlogged.demo.utils.ReferralUtils.generateReferralCode - normal", TestResult.PASS),
                 ReplayTest("ClassUserController.setAndGetUser - recursive setter bug freq logging", TestResult.PASS)
-            ]
-        )
+            ])
+        ),
         # Target(
         # 	"https://github.com/unloggedio/unlogged-spring-webflux-maven-demo",
         # 	"unlogged-spring-webflux-maven-demo",
@@ -496,30 +514,31 @@ if __name__ == "__main__":
         #         ReplayTest("Orchestrator.statusHandler - normal", TestResult.PASS)
         # 	]
         # ),
-        # Target(
-        # 	"https://github.com/unloggedio/unlogged-spring-mvc-maven-demo",
-        # 	"unlogged-spring-mvc-maven-demo",
-        # 	"/pom.xml",
-        # 	"/src/main/java/org/unlogged/mvc/demo/Application.java",
-        # 	buildSystem.MAVEN,
-        # 	[
-        # 		ReplayTest("loadAllBooks - normal", TestResult.PASS),
-        # 		ReplayTest("getBookById - normal", TestResult.PASS),
-        # 		ReplayTest("getAllBooks - normal", TestResult.PASS),
-        # 		ReplayTest("insertBook - normal", TestResult.PASS),
-        # 		ReplayTest("updateBook - normal", TestResult.PASS),
-        # 		ReplayTest("deleteBook - normal", TestResult.PASS)
-        # 	]
-        # )
+        Target(
+        	"https://github.com/unloggedio/unlogged-spring-mvc-maven-demo",
+        	"unlogged-spring-mvc-maven-demo",
+        	"/pom.xml",
+        	"/src/main/java/org/unlogged/mvc/demo/Application.java",
+        	buildSystem.MAVEN,
+            target_run_properties = TargetRunProperties("main","17", StartMode.CMD),
+            replay_test_options= ReplayTestOptions(
+        	[
+        		ReplayTest("loadAllBooks - normal", TestResult.PASS),
+        		ReplayTest("getBookById - normal", TestResult.PASS),
+        		ReplayTest("getAllBooks - normal", TestResult.PASS),
+        		ReplayTest("insertBook - normal", TestResult.PASS),
+        		ReplayTest("updateBook - normal", TestResult.PASS),
+        		ReplayTest("deleteBook - normal", TestResult.PASS)
+        	])
+        )
     ]
 
     passing = True
     result_maps = []
     report_generator = Report_Generator(ReportType.REPLAY)
-
     for local_target in target_list:
         result_map = replay_target(local_target)
-        report_generator.add_replay_result_status_entry(local_target,result_map)
+        report_generator.add_replay_result_status_entry(local_target, result_map)
         result_maps.append(result_map)
         if result_map['status'] == TestResult.FAIL:
             passing = False
